net/libp2p: Split services into separate files
